fix: move search clear button outside of search facets
